+++ v3/deprecated/20250409-v3-reclamm-pool/output/arbitrum.json
@@ -1,0 +1,4 @@
+{
+  "ReClammPoolFactory": "0xa1D0791a41318c775707C56eAe247AF81a05322C",
+  "MockReClammPool": "0xe482ef529B8f7fFC6ECCDea7eaF29C717E9ed079"
+}
