+++ v3/deprecated/20250409-v3-reclamm-pool/output/avalanche.json
@@ -1,0 +1,4 @@
+{
+  "ReClammPoolFactory": "0x1CD776897ef4f647bf8241Ec69549e4A9cb1D608",
+  "MockReClammPool": "0x33608014e176139e0b9058B32EEae1A5ce795FD3"
+}
