+++ v3/deprecated/20250409-v3-reclamm-pool/output/gnosis.json
@@ -1,0 +1,4 @@
+{
+  "ReClammPoolFactory": "0x1780e8185199DAc45284039Bb3E3B64060D943AB",
+  "MockReClammPool": "0x02B4a4b0db94db44cb91290b2b5Eb5c5a8a1Cbc9"
+}
