+++ v3/deprecated/20250409-v3-reclamm-pool/output/mainnet.json
@@ -1,0 +1,4 @@
+{
+  "ReClammPoolFactory": "0x5939ab16fDf1991B0EF603c639B6b501A7841fAB",
+  "MockReClammPool": "0x6A5127AE6df32f722e34B5b4aBc67b0d4e582B39"
+}
