--- v3/tasks/20250409-v3-reclamm-pool/output/base.json
@@ -1,4 +1,0 @@
-{
-  "ReClammPoolFactory": "0x3BEb058DE1A25dd24223fd9e1796df8589429AcE",
-  "MockReClammPool": "0x39110AD9A2Dbfe46c8674693838D8BA95ea1e7A9"
-}
