--- v3/tasks/20250409-v3-reclamm-pool/output/hyperevm.json
@@ -1,4 +1,0 @@
-{
-  "ReClammPoolFactory": "0x25EA7b202CD8447e8E6F4A7C59b19C66DfEf1925",
-  "MockReClammPool": "0x493E0E14fCCB89925b507322f605b84629A8918a"
-}
