--- v3/tasks/20250409-v3-reclamm-pool/output/optimism.json
@@ -1,4 +1,0 @@
-{
-  "ReClammPoolFactory": "0x0b11209B8c5E821b18dED147583b8978c3E63911",
-  "MockReClammPool": "0xaC1675219FbB81F2C53950a92dF3F95AAffeab8E"
-}
