--- v3/tasks/20250409-v3-reclamm-pool/output/sepolia.json
@@ -1,4 +1,0 @@
-{
-  "ReClammPoolFactory": "0x2A4054779DaDf4Cf3F4849f1AC307845F182cfE5",
-  "MockReClammPool": "0x919094AdC27c2B779F70eD65D5d96F0b5331f660"
-}
