--- v3/tasks/20250409-v3-reclamm-pool/readme.md
@@ -1,17 +1,0 @@
-# 2025-04-09 - V3 Readjusting Concentrated Liquidity AMM (ReClAMM)
-
-Deployment of `ReClammPoolFactory`, a new pool type that implements concentrated liquidity by imposing price bounds on the constant product "weighted" math curve (using virtual balances). Current approaches to CL require active management of the position by the user: either by adjusting "ticks" in a Uniswap-style non-fungible position, or by adding/removing liquidity with fungible approaches. The ReClamm uses algorithmic virtual balance modifications to internally and automatically adjust the price interval as needed, 
-
-## Useful Files
-
-- [Code](https://github.com/balancer/reclamm/commit/61512d4737a69ca5b703842d058252585cfca381)
-- [Ethereum mainnet addresses](./output/mainnet.json)
-- [Gnosis mainnet addresses](./output/gnosis.json)
-- [Arbitrum mainnet addresses](./output/arbitrum.json)
-- [Base mainnet addresses](./output/base.json)
-- [Optimism mainnet addresses](./output/optimism.json)
-- [Avalanche mainnet addresses](./output/avalanche.json)
-- [Hyperevm mainnet addresses](./output/hyperevm.json)
-- [Sepolia testnet addresses](./output/sepolia.json)
-- [`ReClammPoolFactory` artifact](./artifact/ReClammPoolFactory.json)
-- [`ReClammPool` artifact](./artifact/ReClammPool.json)
